+++ engine/src/main/java/dev/gamekit/core/Renderer.java
@@ -1,6 +1,7 @@
 package dev.gamekit.core;
 
 import dev.gamekit.graphics.*;
+import dev.gamekit.settings.ImageInterpolation;
 
 import java.awt.*;
 import java.awt.image.BufferedImage;
@@ -112,9 +113,24 @@ public final class Renderer {
     return call;
   }
 
-  /** Draws a scaled <b>center-origin</b> {@link BufferedImage} at (x, y) with size and height */
+  /**
+   * Draws a scaled <b>center-origin</b> {@link BufferedImage} at {@code (x, y)} with width and
+   * height
+   */
   public static DrawImage drawImage(BufferedImage img, int x, int y, int width, int height) {
     DrawImage call = new DrawImage(img, x, y, width, height);
+    BUFFER.add(call);
+    return call;
+  }
+
+  /**
+   * Draws a scaled <b>center-origin</b> {@link BufferedImage} at {@code (x, y)} with size,
+   * height and an {@link ImageInterpolation interpolation}
+   */
+  public static DrawImage drawImage(
+    BufferedImage img, int x, int y, int width, int height, ImageInterpolation interpolation
+  ) {
+    DrawImage call = new DrawImage(img, x, y, width, height, interpolation);
     BUFFER.add(call);
     return call;
   }