+++ engine/src/main/java/dev/gamekit/graphics/DrawImage.java
@@ -1,5 +1,7 @@
 package dev.gamekit.graphics;
 
+import dev.gamekit.settings.ImageInterpolation;
+
 import java.awt.*;
 import java.awt.image.BufferedImage;
 
@@ -9,19 +11,33 @@ public class DrawImage extends DrawCall<DrawImage> {
   private final int x, y;
   private final int width;
   private final int height;
+  private final ImageInterpolation interpolation;
 
-  public DrawImage(BufferedImage image, int x, int y, int width, int height) {
+  public DrawImage(
+    BufferedImage image, int x, int y, int width, int height, ImageInterpolation interpolation
+  ) {
     this.image = image;
     this.x = x;
     this.y = y;
     this.width = width;
     this.height = height;
+    this.interpolation = interpolation;
+  }
+
+  public DrawImage(BufferedImage image, int x, int y, int width, int height) {
+    this(image, x, y, width, height, ImageInterpolation.DEFAULT);
   }
 
   @Override
   protected void draw(Graphics2D g) {
+    ImageInterpolation originalInterpolation = ImageInterpolation.from(g);
+    interpolation.apply(g);
+
     int x0 = x - width / 2, y0 = y + height / 2;
     int x1 = x0 + width, y1 = y0 - height;
     g.drawImage(image, x0, -y0, x1, -y1, 0, 0, image.getWidth(), image.getHeight(), null);
+
+    if (originalInterpolation != null)
+      originalInterpolation.apply(g);
   }
 }