+++ CHANGELOG.md
@@ -31,6 +31,9 @@
 - Added `static void drawPolygon(int[])` and `static void fillPolygon(int[])` methods to
   `Renderer` class
 - Added `Sprite` which is a `Component` that renders an image appearance for an entity
+- Added interpolation support to `DrawImage` draw call
+- Added `void drawImage(BufferedImage, int, int, int, int, ImageInterpolation)` method overload
+  to `Renderer` class
 
 ### Changed